+++ src/Serilog.Sinks.Elasticsearch/Sinks/ElasticSearch/ElasticSearchLogShipper.cs
@@ -22,6 +22,9 @@ using System.Text;
 using System.Threading;
 using Elasticsearch.Net;
 using Serilog.Debugging;
+#if NO_TIMER
+using Serilog.Sinks.Elasticsearch.CrossPlatform;
+#endif
 
 namespace Serilog.Sinks.Elasticsearch
 {
@@ -30,7 +33,11 @@ namespace Serilog.Sinks.Elasticsearch
         private readonly ElasticsearchSinkState _state;
 
         readonly int _batchPostingLimit;
+#if NO_TIMER
+        readonly PortableTimer _timer;
+#else
         readonly Timer _timer;
+#endif
 
         readonly ExponentialBackoffConnectionSchedule _connectionSchedule;
         readonly object _stateLock = new object();
@@ -51,17 +58,12 @@ namespace Serilog.Sinks.Elasticsearch
             _logFolder = Path.GetDirectoryName(_bookmarkFilename);
             _candidateSearchPath = Path.GetFileName(_state.Options.BufferBaseFilename) + "*.json";
 
-            _timer = new Timer(s => OnTick());
-
-            AppDomain.CurrentDomain.DomainUnload += OnAppDomainUnloading;
-            AppDomain.CurrentDomain.ProcessExit += OnAppDomainUnloading;
-
+#if NO_TIMER
+            _timer = new PortableTimer(cancel => OnTick());
+#else
+            _timer = new Timer(s => OnTick(), null, -1, -1);
+#endif
             SetTimer();
-        }
-
-        void OnAppDomainUnloading(object sender, EventArgs args)
-        {
-            CloseAndFlush();
         }
 
         void CloseAndFlush()
@@ -74,12 +76,15 @@ namespace Serilog.Sinks.Elasticsearch
                 _unloading = true;
             }
 
-            AppDomain.CurrentDomain.DomainUnload -= OnAppDomainUnloading;
-            AppDomain.CurrentDomain.ProcessExit -= OnAppDomainUnloading;
+
+#if NO_TIMER
+            _timer.Dispose();
+#else
 
             var wh = new ManualResetEvent(false);
             if (_timer.Dispose(wh))
                 wh.WaitOne();
+#endif
 
             OnTick();
         }
@@ -108,8 +113,11 @@ namespace Serilog.Sinks.Elasticsearch
         {
             // Note, called under _stateLock
             var infiniteTimespan = Timeout.InfiniteTimeSpan;
-
+#if NO_TIMER
+            _timer.Start(_connectionSchedule.NextInterval);
+#else
             _timer.Change(_connectionSchedule.NextInterval, infiniteTimespan);
+#endif
         }
 
         void OnTick()
@@ -130,7 +138,7 @@ namespace Serilog.Sinks.Elasticsearch
                     // Locking the bookmark ensures that though there may be multiple instances of this
                     // class running, only one will ship logs at a time.
 
-                    using (var bookmark = File.Open(_bookmarkFilename, FileMode.OpenOrCreate, FileAccess.ReadWrite, FileShare.Read))
+                    using (var bookmark = System.IO.File.Open(_bookmarkFilename, FileMode.OpenOrCreate, FileAccess.ReadWrite, FileShare.Read))
                     {
                         long nextLineBeginsAtOffset;
                         string currentFilePath;
@@ -139,7 +147,7 @@ namespace Serilog.Sinks.Elasticsearch
 
                         var fileSet = GetFileSet();
 
-                        if (currentFilePath == null || !File.Exists(currentFilePath))
+                        if (currentFilePath == null || !System.IO.File.Exists(currentFilePath))
                         {
                             nextLineBeginsAtOffset = 0;
                             currentFilePath = fileSet.FirstOrDefault();
@@ -164,7 +172,7 @@ namespace Serilog.Sinks.Elasticsearch
 
                         var payload = new List<string>();
 
-                        using (var current = File.Open(currentFilePath, FileMode.Open, FileAccess.Read, FileShare.ReadWrite))
+                        using (var current = System.IO.File.Open(currentFilePath, FileMode.Open, FileAccess.Read, FileShare.ReadWrite))
                         {
                             current.Position = nextLineBeginsAtOffset;
 
@@ -215,7 +223,7 @@ namespace Serilog.Sinks.Elasticsearch
                                 // best to move on, though a lock on the current file
                                 // will delay this.
 
-                                File.Delete(fileSet[0]);
+                                System.IO.File.Delete(fileSet[0]);
                             }
                         }
                     }
@@ -243,7 +251,7 @@ namespace Serilog.Sinks.Elasticsearch
         {
             try
             {
-                using (var fileStream = File.Open(file, FileMode.OpenOrCreate, FileAccess.ReadWrite, FileShare.Read))
+                using (var fileStream = System.IO.File.Open(file, FileMode.OpenOrCreate, FileAccess.ReadWrite, FileShare.Read))
                 {
                     return fileStream.Length <= maxLen;
                 }